requirements.txt
certifi==2022.9.14
coloredlogs==15.0.1
docker==6.0.0
python_dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
flask-apscheduler==1.12.4
flask==2.2.2